requirements.txt
Django==2.0
djangorestframework==3.7.3
inflect==0.2.5
mysqlclient==1.3.12
pytz==2017.3
